MobX: Tout début refonte 'component'
